18/4

+++ Fontend_ReactJS/react_admin/src/Function.js
@@ -1,7 +1,39 @@
 import api from './api'
 export default {
     getData: async (data) => {
-        const response = await api.getData(data)
+        const response = await api.getData(Object.assign({token: localStorage.getItem("accessToken")},data))
+        if (response && response.data) {
+            return response.data;
+        }
+        return;
+    },
+
+    getAllData: async (data) => {
+        const response = await api.getAllData(Object.assign({token: localStorage.getItem("accessToken")},data))
+        if (response && response.data) {
+            return response.data;
+        }
+        return;
+    },
+
+    getTableFromID: async (data) => {
+        const response = await api.getTableFromID(Object.assign({token: localStorage.getItem("accessToken")},data))
+        if (response && response.data) {
+            return response.data;
+        }
+        return;
+    },
+
+    getEmployeeFromToken: async () => {
+        const response = await api.getEmployeeFromToken({token: localStorage.getItem("accessToken")})
+        if (response && response.data) {
+            return response.data;
+        }
+        return;
+    },
+
+    getKHFromMKH: async (data) => {
+        const response = await api.getKHFromMKH(data)
         if (response && response.data) {
             return response.data;
         }
@@ -9,7 +41,7 @@ export default {
     },
 
     postData: async (data) => {
-        const response = await api.postData(data)
+        const response = await api.postData(Object.assign({token: localStorage.getItem("accessToken")},data))
         if (response && response.data) {
             return response;
         }
@@ -17,7 +49,7 @@ export default {
     },
 
     editData: async (data) => {
-        const response = await api.editData(data)
+        const response = await api.editData(Object.assign({token: localStorage.getItem("accessToken")},data))
         if (response && response.data) {
             return response;
         }
@@ -25,7 +57,7 @@ export default {
     },
 
     deleteData: async (data) => {
-        const response = await api.deleteData(data)
+        const response = await api.deleteData(Object.assign({token: localStorage.getItem("accessToken")},data))
         if (response && response.data) {
             return response;
         }
@@ -33,7 +65,7 @@ export default {
     },
 
     detailData: async (data) => {
-        const response = await api.detailData(data)
+        const response = await api.detailData(Object.assign({token: localStorage.getItem("accessToken")},data))
         if (response && response.data) {
             return response;
         }
@@ -41,12 +73,11 @@ export default {
     },
 
     changeText: (str) => {
-        var convertToArray = str.toLowerCase().split(' ');
-        var result = convertToArray.map(function (val) {
-            return val.replace(val.charAt(0), val.charAt(0).toUpperCase());
-        });
-
-        return result.join(' ');
+        str = str.toString();
+        let words = str.split(" ").map(word => {
+            return word[0].toUpperCase() + word.slice(1);
+        })
+        return words.join(" ");
     },
 
 }
